Changed implementation of `next`
Most previous commits utilized my invented implementation of `next`. The
implementation presented by the author was much cleaner so I chose to
change to a more similar implemetation.

Notice that we now have an acceptable version of our core logic with no
setup, no mocks and no stubs. This structure is a consequence of
utilizing pure functions. We have not even added React or Redux.