Change `core.vote` to only handle `vote` substate
Previous commits of the `core` functions expected to receive the
*entire* application state. This application is small enough that
handling the entire state is not too complicated. However, as a general
rule, the code is more maintainable if these `core` functions only
handle the state in which they are interested.

This change not only affects the code in `core` but the code in
`reducer`. `reducer` now has an additional responsibility of extracting
the substate that `core.vote` handles.