Correct error when single vote
Previous commits, when a single vote was cast and it was for the first
entry in a pair, incorrectly selected the *second* entry as the winner.

This incorrect behavior was caused by `Map.get` and `Map.getIn`
returning `null` when the key sought was *not* present in the `Map`. The
correction involved supplying a default value for `get` and `getIn`.